refactor(test) : 숫자 배열을 받아서 List<Ball>을 리턴하는 헬퍼 메소드 구현
- 숫자의 갯수에 맞게 Ball을 가지고 있는 리스트를 리턴하여 테스트 코드 리팩토링
