+++ src/test/java/domain/LottosTest.java
@@ -4,17 +4,18 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
+import java.util.HashMap;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class LottoStatisticsTest {
-
-    private LottoStatistics lottoStatistics;
+public class LottosTest {
+    private Lottos lottos;
+    private WinningLotto winningLotto;
 
     @BeforeEach
     public void createLottos() {
-        WinningLotto winningLotto = new WinningLotto(LottoTest.getLottoBalls(1,7,8,9,10,11),34);
-        Lottos lottos = new Lottos(
+        winningLotto = new WinningLotto(LottoTest.getLottoBalls(1,7,8,9,10,11),34);
+        lottos = new Lottos(
                 Arrays.asList(
                         new Lotto(LottoTest.getLottoBalls(1,7,8,9,10,11)),
                         new Lotto(LottoTest.getLottoBalls(1,7,8,9,10,11)),
@@ -24,18 +25,12 @@ public class LottoStatisticsTest {
                         new Lotto(LottoTest.getLottoBalls(1,7,8,9,10,13))
                 )
         );
-        lottoStatistics = new LottoStatistics(lottos.getRankCounts(winningLotto), lottos.getLottos().size());
     }
 
     @Test
-    public void calculateTotalRewardTest() {
-        assertThat(lottoStatistics.calculateTotalReward()).isEqualTo(1709532704);
-        System.out.println(lottoStatistics.toString());
-    }
-
-    @Test
-    public void calculateEarningsRateTest() {
-        assertThat(lottoStatistics.calculateEarningsRate(6000)).isEqualTo(284922.0);
-        assertThat(lottoStatistics.calculateEarningsRate(6061500000L)).isEqualTo(0.0);
+    void getRankCounts(){
+        HashMap<LottoRank, Integer> rankCounts = lottos.getRankCounts(winningLotto);
+        assertThat(rankCounts.get(LottoRank.FIRST)).isEqualTo(3);
+        assertThat(rankCounts.get(LottoRank.THIRD)).isEqualTo(3);
     }
 }
